X86: fixed a locale issue/ naming convention applied to x86 image scripts

+++ scripts/x86config.sh
@@ -77,7 +77,11 @@ echo "  Installing grub-efi-amd64 to make the 64bit UEFI bootloader"
 
 apt-get update
 apt-get -y install grub-efi-amd64-bin
-grub-mkstandalone --compress=gz -O x86_64-efi -o /boot/efi/BOOT/BOOTX64.EFI -d /usr/lib/grub/x86_64-efi --modules="part_gpt part_msdos" --fonts="unicode" --locales="en@quot" --themes="" /boot/grub/grub.cfg
+grub-mkstandalone --compress=gz -O x86_64-efi -o /boot/efi/BOOT/BOOTX64.EFI -d /usr/lib/grub/x86_64-efi --modules="part_gpt part_msdos" --fonts="unicode" --themes="" /boot/grub/grub.cfg
+if [ ! -e /boot/efi/BOOT/BOOTX64.EFI ]; then
+	echo "Fatal error, no 64bit bootmanager created, aborting..." 
+    exit 1
+fi
 #we cannot install grub-efi-amd64 and grub-efi-ia32 on the same machine.
 #on the off-chance that we need a 32bit bootloader, we remove amd64 and install ia32 to generate one
 echo "  Uninstalling grub-efi-amd64"
@@ -85,7 +89,11 @@ apt-get -y --purge remove grub-efi-amd64-bin
 
 echo "  Installing grub-efi-ia32 to make the 32bit UEFI bootloader"
 apt-get -y install grub-efi-ia32-bin
-grub-mkstandalone --compress=gz -O i386-efi -o /boot/efi/BOOT/BOOTIA32.EFI -d /usr/lib/grub/i386-efi --modules="part_gpt part_msdos" --fonts="unicode" --locales="en@quot" --themes="" /boot/grub/grub.cfg
+grub-mkstandalone --compress=gz -O i386-efi -o /boot/efi/BOOT/BOOTIA32.EFI -d /usr/lib/grub/i386-efi --modules="part_gpt part_msdos" --fonts="unicode" --themes="" /boot/grub/grub.cfg 
+if [ ! -e /boot/efi/BOOT/BOOTIA32.EFI ]; then
+	echo "Fatal error, no 32bit bootmanager created, aborting..." 
+    exit 1
+fi
 #and remove it again
 echo "  Uninstalling grub-efi-ia32-bin"
 apt-get -y --purge remove grub-efi-ia32-bin
